Improve http request cancellation
Switches the Promise.race to an AbortController, which means that we're
using the recommended way to abort fetch requests.

https://github.com/node-fetch/node-fetch#request-cancellation-with-abortsignal

The old way actually led to the process waiting for the HTTP request to
finish anyway (not really a problem for Lambda, but good praxis). This
is because hanging promises (which Promise.race is wont to do) will keep
the node process running until they're done.

This seemed to change around about node v16. Can't remember exactly
where.